Add NHT warp reduction for evalBackwardNoKBuffer path
In the k_buffer_size=0 backward path, all 32 warp threads process the
same particle sequentially. Replace per-thread atomicAdds to the global
feature gradient buffer with a two-step warp reduction:

1. featuresIntegrateBwdToLocalGrad: write feature grads into a
   thread-private float[ParticleFeatureDim] buffer using the pointer
   offset trick (featureLocalGrad - particleOffset passed as gradPtr
   with exclusiveGradient=true), avoiding any atomics.

2. featureLocalGradWarpReduceAndWrite: __shfl_xor_sync butterfly reduce
   across all 32 lanes, then lane 0 atomicAdds the warp sum to the
   global gradient buffer. Reduces atomics from 32×ParticleFeatureDim
   to ParticleFeatureDim per particle per warp.

Also fix fused_ssim to cast inputs to float32 for fp16 feature_output_half
compatibility, and improve slangc invocation to use a tmp file to avoid
spurious ninja rebuilds when Slang output is unchanged.

+++ threedgrut/model/losses.py
@@ -30,4 +30,5 @@ def l2_loss(network_output, gt):
 @torch.cuda.nvtx.range("ssim")
 def ssim(img1, img2, window_size=11, size_average=True):
     # predicted_image, gt_image: [BS, CH, H, W], predicted_image is differentiable
-    return fused_ssim(img1, img2, padding="valid")
+    # fused_ssim requires float32; cast here to handle fp16 feature_output_half=true paths
+    return fused_ssim(img1.float(), img2.float(), padding="valid")
